Ignoring MAKEFLAGS in rev. 1.15 was a very bad idea.  This causes wrong share/mk files to be used, which can be fatal with upgrades or downgrades, e.g., when building RELENG_5 on HEAD.
Reported by:	glebius

For now, just exclude -P from MAKEFLAGS when running crunchgen(1).
(Note that it will still break when run with certain -d options.)

The real solution is to fix make(1) to not print stuff on stdout
when it's not supposed to, e.g., through the -P and -dX options,
and to fix crunchgen(1) to not redirect stderr to stdout when
running make(1).  Once this is implemented, this hack can go.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@140508 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f