Update zconf.h and gzio.c with changes from ZLib 1.2.1 to ZLib 1.2.2.
These are two of the three files that have non-trivial differences from
the vendor branch.  minigzip.c is the third, but there were no changes
from ZLib 1.2.1 to ZLib 1.2.2 in that file.

The rest of the files I intend to get reverted back to the vendor
branch (with cooperation of cvsadmin@).

PR:		i386/76294


git-svn-id: file:///home/git/svn/freebsd_mirror/head@145490 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f