For non-profiling kernels, there were two symbols assigned to the same address.  One was alltraps_with_regs_pushed, the other was calltrap.
When the stack tracer walks up, it looks for magic symbol names to
determine how to parse non-standard stack frames, such as a trapframe.
It was looking for "calltrap".  Which of the two symbols you got depended
on things like Phase of moon, etc.  If you were unlucky, you got a
garbage stack trace for things like 'debug.trace_on_panic', which would
completely hide the actual source of the problem.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@146461 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f