Change hw.pci.do_powerstate from a boolean to a range.  0 means the same as today: do no power management.  1 means be conservative about what you power down (any device class that has caused problems gets added here).  2 means be agressive about what gets powered down (any device class that's fundamental to the system is here).  3 means power them all down, reguardless.  The default is 1.
The effect in the default system is to add mass storage devices to the
list that we don't power down.  From all the pciconf -l lists that
I've seen for the aac and amr issue, the bad device has been a mass
storage device class.

This is an attempt at a compromise between the very small number of
systems that have extreme issues with powerdown, and the very large
number of systems that gain real benefits from powerdown (I get about
20% more battery life when I attach a minimal set of drivers on my
Sony).  Hopefully it will strike the proper balance.

MFC After: 3 days (before next beta)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@149974 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f