Comment test_lseek().
Add test_kqueue(), which registers and unregisters various kqueue filter
types on a fifo in order to make sure that EVFILT_READ, EVFILT_WRITE can
be registered, and that EVFILT_NETDEV can't be registered.  For now, we
don't test that EVFILT_VNODE can be registered on fifos, as that has been
broken at some point.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@150064 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f