It is unclear who is wrong and who is right, but when operating on plain file bsdlabel(8) always writes label at a fixed offset from its beginning (512 bytes), regardless of the sector size. At the same time, bsdlabel geom class expects label to be available at the very beginning of the second sector.
As a result, images prepared in userland for media with sector size
different from 512 bytes (i.e. 2k for cdroms) are not recognized by
the tasting mechanism.

Solve the problem by always looking for the label at 512-byte offset
if we can't find it at the beginning of the second sector and sector
size is not 512 bytes.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@152972 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f