Remove broken code that mucks about with tcsetpgrp() -- even if su isn't the foreground process.  Hopefully this won't break PAM, but I couldn't find any useful information about ache's theory that it will.
Specifically, this change fixes the following:

    # sh
    # echo $$
    # su - root -c id &
    # echo $$

The PID output changes as su seems to be kill -STOP'ing itself
and catching the parent shell in the process.  This is especially
bad if you add a ``su - user -c command &'' to an rc script!

Sponsored by:		Sophos/Activestate
Not objected to by:	des


git-svn-id: file:///home/git/svn/freebsd_mirror/head@153964 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f