Allow root in the host environment to open ptys within jailed environments. This logic change was introduced in revision 1.74:
Correct an oversight in jail() that allowed processes in jail to access
ptys in ways that might be unethical, especially towards processes not in
jail, or in other jails.

It should be fine to allow root in the host environment to do this. This
allows for more effective monitoring of prisons from the host environment.

Discussed with:	rwatson
MFC after:	1 week


git-svn-id: file:///home/git/svn/freebsd_mirror/head@155117 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f