Fix "slow (on-the-fly) zero fills percentage (`%slo-z')" some more.  The value printed is actually the optimized (i.e., the non-slow, not-on-the-fly zero fills percentage) except in overflow cases.  Describe it as %ozfod in the display.  Move the field descriptor 1 to the left so that there is space for 5 characters after the % sign (this leaves no space between the number and the descriptor but the % character serves well as a separator).
Fixed integer overflow at z.ozfod = UINT_MAX/100 in the calculation of
%ozfod.  This value can be reached just a few hours or minutes after
booting, so %ozfod was usually garbage in boot mode.  Now %ozfod is
correct in boot mode for a few days or hours.

Print a non-dummy %ozfod when the division for it isn't division by 0
instead of when the result will be less than 100%.  A result of 100%
may be correct, though a result of more than 100% indicates overflow
of one or both counters.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@158172 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f