A small race existed where the lock was dropped between when encif was tested and then set. [1]
Reorganise things to eliminate this, we now ensure that enc0 can not be
destroyed which as the benefit of no longer needing to lock in
ipsec_filter and ipsec_bpf. The cloner will create one interface during the
init so we can guarantee that encif will be valid before any SPD entries are
added to ipsec.

Spotted by:	glebius [1]


git-svn-id: file:///home/git/svn/freebsd_mirror/head@160011 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f