Teach mount(8) about a 'late' keyword, which means the file system should not be mounted unless the -l flag was specified.
Add an rc script, mountlate, which basically runs 'mount -a -l'.  It runs
after DAEMON but before LOGIN.

This is useful for things like loopback mounts, because mountcritremote
runs before mountd  / nfsd (since /usr might be a remote file system), so
an attempt to mount a loopback network file system in mountcritremote will
fail.

Also add a progress message to mountcritlocal, for the sake of symmetry
with similar messages in mountcritremote and mountlate.

Reviewed by:	freebsd-rc
MFC after:	3 weeks


git-svn-id: file:///home/git/svn/freebsd_mirror/head@160303 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f