Change semantics of socket close and detach.  Add a new protocol switch function, pru_close, to notify protocols that the file descriptor or other consumer of a socket is closing the socket.  pru_abort is now a notification of close also, and no longer detaches.  pru_detach is no longer used to notify of close, and will be called during socket tear-down by sofree() when all references to a socket evaporate after an earlier call to abort or close the socket.  This means detach is now an unconditional teardown of a socket, whereas previously sockets could persist after detach of the protocol retained a reference.
This faciliates sharing mutexes between layers of the network stack as
the mutex is required during the checking and removal of references at
the head of sofree().  With this change, pru_detach can now assume that
the mutex will no longer be required by the socket layer after
completion, whereas before this was not necessarily true.

Reviewed by:	gnn


git-svn-id: file:///home/git/svn/freebsd_mirror/head@160549 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f