Introduce a new method ipv6if which attemptes to figure out if an interface is an IPv6 interface.
Use this method to decide if we should attempt to configure an interface
with an IPv6 address in pccard_ether.  The mechanism pccard_ether uses
to do this is unsuited to the task because it assumes the list of
interfaces it is passed is the full list of IPv6 interfaces and makes
decissions based on that.  This is at least a step in the right
direction and is probably about as much as we can MFC safely.

PR:		conf/103428
MFC after:	3 days


git-svn-id: file:///home/git/svn/freebsd_mirror/head@162490 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f