MFp4: 	o Fix the packet statistics 	o Make sure we set the FD bit when in full duplex 	o Improve TX side efficency by eliminating a data copy for 	  unfragmented mbufs (the hardware can't do s/g). 	o Minor busdma pedantry 	o better comments in some places, more XXX in others 	o Minor style nits.
This solves a problem I was seeing where I'd get no ethernet when not
booting with a NFS root.  Well, unless I unplugged the cable and
plugged it back in first so I'd get the same up down up messages I get
for NFS root...

Thanks to sam and scottl for suggestions on making this driver more
efficient through better use of approrpiate APIs.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@163937 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f