Set channel utility for Hauuapuge PVR-250 and PVR-350.
This s part of an import of the PVR-250 driver.  Originally it was
calleed pvr250-setchannel, but it seems better to improve this program
to work for any tuner card, so I'm starting with a more generic name.
That shouldn't mislead anybody: currently the program only works with
the (yet to be committed) cxm driver.

Contributed by: John Wehle <john\@feith.com>


git-svn-id: file:///home/git/svn/freebsd_mirror/head@165023 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f