Cleanup of netatalk atalk layer includes, protocol definitions, and routing:
- style(9) cleanup -- white space, braces, etc.

- Make include guards consistent with our more general naming
  convention.

- Rearrange and complete forward structure declarations in at_extern.h,
  remove testing of guards of various other include files to protect
  function declarations.

This leaves an ifdef _KERNEL in at_var.h, but from inspection it seems
likely that this file is not actually safe for inclusion in user space
still.  However, since it's not included from within src/ so this does
not appear to be an issue (ifconfig, etc, have migrated to the generic
cross-protocol ioctls for address operations).


git-svn-id: file:///home/git/svn/freebsd_mirror/head@165972 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f