Disable RPC exponential back-off for FreeBSD.org systems (IE. hidden behind _FREEFALL_CONFIG).  This is done mainly to make NIS even more resistant to packet loss.
This is not enabled by default for "normal" FreeBSD since it might cause
the server providing the RPC service to be hit heavily with RPC traffic
in case of problems.  freefall.FreeBSD.org and hub.FreeBSD.org have been
running with a patch similar to this for a couple of weeks.

MFC after:	1 week
Discussed with:	peter


git-svn-id: file:///home/git/svn/freebsd_mirror/head@167199 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f