Better fix for the errors under high load.  Returning CAM_SCSI_BUSY is almost never correct as CAM has no real understanding of it, and will just immediately retry the command.  This leads to undesirable cycling of the camisr as well as a high possibility for the command to exhaust its retries before the driver can get around to servicing it.
The better fix, as demonstrated here, is to freeze the simq and mark the
command as needing to be tried.  Then when driver can service the command,
the simq gets unfrozen.  This is correct, and documented here to help reduce
the mystery.  However, it also points out a shortcoming in CAM error handling
that makes writing drivers harder.

Submitted by: Erich Chen


git-svn-id: file:///home/git/svn/freebsd_mirror/head@167254 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f