Add a pn_destroy field to pfs_node.  This field points to a destructor function which is called from pfs_destroy() before the node is reclaimed.
Modify pfs_create_{dir,file,link}() to accept a pointer to a destructor
function in addition to the usual attr / fill / vis pointers.

This breaks both the programming and binary interfaces between pseudofs
and its consumers.  It is believed that there are no pseudofs consumers
outside the source tree, so that the impact of this change is minimal.

Submitted by:	Aniruddha Bohra <bohra@cs.rutgers.edu>


git-svn-id: file:///home/git/svn/freebsd_mirror/head@167482 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f