Simplifly syncache_expand() and clarify its semantics.  Zero is returned when the ACK is invalid and doesn't belong to any registered connection, either in syncache or through SYN cookies.  True but a NULL struct socket is returned when the 3WHS completed but the socket could not be created due to insufficient resources or limits reached.
For both cases an RST is sent back in tcp_input().

A logic error leading to a panic is fixed where syncache_expand() would
free the mbuf on socket allocation failure but tcp_input() later supplies
it to tcp_dropwithreset() to issue a RST to the peer.

Reported by:	kris (the panic)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@168902 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f