Make suid/sgid restore be "opportunistic" if owner restore is not requested.  If you ask for permissions to be restored but not owner, you will now get no error if suid/sgid bits cannot be set.  (It's a security hole to restore suid/sgid bits if the owner/group aren't restored.)
This fixes an obscure problem where a simple
"tar -xf" with no other options will sometimes
fail gratuitously because of suid/sgid bits.
This is causing occasional problems for people
using bsdtar as a drop-in replacement for
"that other tar program." ;-)

Note: If you do ask for owner restore, then suid/sgid
restore failures still issue an error.  This
only suppresses the error in the case where an
suid/sgid bit restore fails because of an owner
mismatch and owner restore was not requested.

Approved by: re (bmah)
MFC after: 7 days


git-svn-id: file:///home/git/svn/freebsd_mirror/head@171811 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f