When moving a directory across devices to a place where a directory with the same name exists, delete that directory first, before performing the copy.  This ensures that mv(1) across devices follows the semantics of rename(2), as required by POSIX.
This change could introduce the potential of data loss, even if the
copy fails, violating the atomicity properties of rename(2).  This is
(mostly) mitigated by first renaming the destination and obliterating
it only after a succesfull copy.

The above logic also led to the introduction of code that will cleanup
the results of a partial copy, if a cross-device copy fails.

PR:		bin/118367
MFC after:	1 month


git-svn-id: file:///home/git/svn/freebsd_mirror/head@174664 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f