Add a /S mode to DDB "ex" command, which interprets and prints the value at the requested address as a symbol.  For example, "ex /S aio_swake" prints the name of the function currently registered in via aio_swake hook.
The change as committed differs slightly from the patch in the PR,
as I force the size of the retrieved value (and the automatic
address increment) to be sizeof(void *).  This seems to provide
the most useful auto-increment behavior, and defaults using the
default size (4), which is not sizeof(void *) on 64-bit platforms.

MFC after:	3 days
PR:		57976
Submitted by:	Dan Strick <strick at covad.net>


git-svn-id: file:///home/git/svn/freebsd_mirror/head@176914 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f