Identify ICH9 USB controllers.
I've taken a slightly different approach than is used with the ICH8 controllers
in that each controller is not identified individually (eg USB A, USB B, etc).
Instead I've given then same description to each one even though the device ID
differs.  This can easily be changed if desired, or ICH8 (and any others using
that approach) can be made to work as this does.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@178099 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f