Change the default value of synchronous_dhclient to NO.
To preserve the existing behavior of etc/rc.d/netif, add code to wait
up to if_up_delay seconds (30 seconds by default) for a default route to
be configured if there are any dhcp interfaces.  This should be extended
to test that the interface is actually up.

X-MFC after:


git-svn-id: file:///home/git/svn/freebsd_mirror/head@179003 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f