This is a rewritten driver for the SMSC LAN91C111.  It's based in part on the sn(4) driver and also looking at newer drivers.  The reason for the rewrite is to support MII and to try and resolve some performance issues found when trying to use the sn(4) driver on the Gumstix network boards.
For reference, the SMSC LAN91C111 is a non-PCI ethernet part whose lineage
dates back to Ye Olde Days of ISA.  It seems to get some use in the embedded
space these days on parts lacking on-board MACs or on-board PCI controllers,
such as the XScale PXA line of ARM CPUs.

This also includes a driver for the SMSC LAN83C183 10/100 PHY.

Man page to follow.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@179592 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f