Support for the XScale PXA255 SoC as found on the Gumstix Basix and Connex boards.  This is enough to net-boot to multiuser.
Also supported is the SMSC LAN91C111 parts used on the netCF, netDUO and netMMC
add-on boards.

I'll be putting some instructions on how to boot this on the Gumstix boards
online soon.

This is still fairly rough and will be refined over time but I felt it was
better to get this out there where other people can help out.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@179595 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f