Make uart(4) the default serial port driver on i386 and amd64.
The uart(4) driver has the advantage of supporting a wider variety of
hardware on a greater amount of platforms. This driver has already been
the standard on platforms such as ia64, powerpc and sparc64.

I've decided not to change anything on pc98. I'd rather let people from
the pc98 team look at this.

Approved by:	philip (mentor), marcel


git-svn-id: file:///home/git/svn/freebsd_mirror/head@180487 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f