Put a global variables, which were virtualized but formerly missed under VIMAGE_GLOBAL.
Start putting the extern declarations of the  virtualized globals
under VIMAGE_GLOBAL as the globals themsevles are already.
This will help by the time when we are going to remove the globals
entirely.

While there garbage collect a few dead externs from ip6_var.h.

Sponsored by:	The FreeBSD Foundation


git-svn-id: file:///home/git/svn/freebsd_mirror/head@185937 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f