Add explicit casting in few places.
It is only really necessary for open(2)'s third argument, which is optional and
obtained through stdarg(3). open(2)'s third argument is 32bit and we pass 64
bits. On little endian it works, because we take lower 32 bits, but on big
endian platforms we take upper 32 bits, so we end up with 0.

Reported by:	Milan Čermák <Milan.Cermak@Sun.COM>


git-svn-id: file:///home/git/svn/freebsd_mirror/head@188934 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f