Extend the "vfsopt" mount options for more general use.  Make struct vfsopt and the vfs_buildopts function public, and add some new fields to struct vfsopt (pos and seen), and new functions vfs_getopt_pos and vfs_opterror.
Further extend the interface to allow reading options from the kernel
in addition to sending them to the kernel, with vfs_setopt and related
functions.

While this allows the "name=value" option interface to be used for more
than just FS mounts (planned use is for jails), it retains the current
"vfsopt" name and <sys/mount.h> requirement.

Approved by:	bz (mentor)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@189290 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f