Improve device mode (gadget) stall handling.
Some hardware easily comes out of sync with regard to whether the current or
the next control transfer should be stalled, if a stall command is always
issued before receiving the SETUP packet. After this patch the stall command
will only be issued when a transfer should actually be stalled.

Submitted by:	Hans Petter Selasky


git-svn-id: file:///home/git/svn/freebsd_mirror/head@192552 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f