Make the proof-of-concept UTF-8 support in Syscons less useless.
Add a small Unicode-to-CP437 remapping table to at least demonstrate
that the terminal emulator is perfectly capable of handling UTF-8. This
will of course break if the user loads a different font map, but it at
least allows people to give it a try.

I can now see the box drawing in dialog(1) and the arrows in mutt(1)
correctly.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@194103 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f