Fix race condition in noclobber option.
Formerly, it was possible for the file to be created between the check if it
existed and the open; the contents would then be lost.

Because this must use O_EXCL, noclobber > will not create a file through a
symlink anymore. This agrees with behaviour of other shells.

Approved by:	ed (mentor) (implicit)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@194560 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f