Remove support for the /dev/net/* per-interface devices.  They serve little purpose and are unused in the base system.
The IOCTL functionality is entirely duplicated and routing sockets
provide a richer interface than the kqueue functionality.

Further, it is not practical for these devices to be made sensible in
the face of VIMAGE.

Bump __FreeBSD_version on the off chance that there is any code out
there that actually uses this stuff.

Reviewed by:	rwatson
Discussed with:	bz, zec
Approved by:	re@ (kensmith)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@195175 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f