Migrate init(8) towards utmpx.
According to a comment, we cannot safely remove utmpx entries here
anymore. This is because the libc routines may block on file locking. In
an ideal world login(1) should just remove the entries, which is why I'm
disabling this code for now. If it turns out we get lots of stale
entries here, we should figure out a way to deal with that.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@202194 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f