Port ftpd to utmpx.
Unfortunately I have to partially wreck its functionality, though. ftpd
used to keep a file descriptor to the wtmp, which allowed it to work
from within a chroot. The current utmpx implementation doesn't offer a
way to do this. Maybe we can address this in the future, if it turns out
to be a real issue.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@202209 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f