Store the mutex in the correct location.  Before, we were storing it in the pcb at the td_lock offset, rather than in the struct thread at the td_lock offset.  And we were storing a pointer to the old thread rather than to the mutex.  Why this didn't always kill us, I'll never know.
Fix an obsolete comment and update the prototype in the comments.
Also note what variables are in what registers since this function is
a little complex...

neel@ found this problem and proposed this fix.  This cures a number
of different problem reports out there, and gets us booting octeon to
the login prompt...

Submitted by:	neel@
Reviewed by:	rrs@, gonzo@


git-svn-id: file:///home/git/svn/freebsd_mirror/head@202983 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f