Create the "cfecons" tty directly using tty_makedev(). It is not clear what the intention of having two ttys pointing to the same cfe console device was.
Also we were not initializing the output[] array passed in as input to
tty_makedev() so one name of the ttys was garbage.

Fix the code that calls cfe_write() to deal with the case where only a partial
buffer is written out.

cfe_cngetc() needs to return if there is no character available as input.
If we don't do this then the cfe_timeout() function will spin forever
because cfe_cngetc() will only ever return if there is valid input.

Approved by: imp (mentor)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@203001 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f