Install termcap.small.
Right now we have a termcap.small in the tree, but we don't install it.
If we do install this file by default, it is more likely for
applications to work in single user mode.

I am not entirely happy with the contents of this file. In my opinion we
should remove the `dosansi', `pc' and `pc3' entries and replace them
with `vt100' in case someone uses a serial console.

The file does already have entries for `cons25' and `xterm', which is
used most often.

Requested by:	brucec


git-svn-id: file:///home/git/svn/freebsd_mirror/head@204820 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f