printf: Remove support for building as a csh builtin.
The #define BUILTIN was for building as a csh (not tcsh) builtin.
Given that csh was replaced by tcsh years ago there is no point in keeping
this.

The #define SHELL is for building as an sh builtin and is in active use.

This commit does not change the /bin/sh and /usr/bin/printf binaries.


git-svn-id: svn://svn.freebsd.org/base//head@216310 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f