Fix setting LUN numbers in the mps(4) driver.
Prior to this change, the addressing method wasn't getting set, and
so the LUN field could be set incorrectly in some instances.

This fix should allow for LUN numbers up to 16777215 (and return an error
for anything larger, which wouldn't fit into the flat addressing model).

Submitted by:	scottl (in part)


git-svn-id: svn://svn.freebsd.org/base//head@216368 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f