Make clang default on x86 and powerpc, but not on other architectures. Make fdt default on arm and powerpc.
This now includes cross compiled targets, where before we tried to
make it host-based.

Also, move the lists of default yes and no options to a variable.

In general, only build tools should get this treatment in bsd.own.mk.
Also, the use of TARGET* in the bsd.*mk files is discouraged, but
necessary here due to the ordering of things in buildworld.  We make
the native case work by testing MACHINE_ARCH after TARGET_ARCH.


git-svn-id: svn://svn.freebsd.org/base/head@220359 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f