Rather than trusting that nothing is going to sneak in before the early_late_divider in the second run (and thus be skipped altogether), keep a list of the scripts run early, and use that list to skip things in the second run.
This has the primary benefit of not skipping a local script that gets
ordered too early in the second run. It also gives an opportunity to
clean up/simplify the code a bit.

Use a space-separated list rather than the more traditional colon for
maximum insurance against creativity in local naming conventions.

Reviewed by:	brooks


git-svn-id: http://svn.freebsd.org/base/head@221227 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f