Implement BIO_DELETE for vnode devices by simply overwriting the deleted sectors with all-zeroes.
The zeroes come from a static buffer; null(4) uses a dynamic buffer for
the same purpose (for /dev/zero).  It might be a good idea to have a
static, shared, read-only all-zeroes page somewhere in the kernel that
md(4), null(4) and any other code that needs zeroes could use.

Reviewed by:	kib
MFC after:	3 weeks


git-svn-id: http://svn.freebsd.org/base/head@221229 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f