The security run requests unmaskable output, even if the only output is to tell that there is a separate email or that the output is logged to a file.
This commit changes the return code for the non-inline case to tell that
this message is not important enough and can be masked if necessary. The
messages from the security checks themself are not affected by this and
show up as before in the periodic security email/file.

The inline case still requests to not mask the output, as with the current
way of handling this there is no easy way to handle this.

PR:			138692
Analysis/patch atch by:	Chris Cowart <ccowart@timesinks.net>
X-MFC after:		on request


git-svn-id: svn://svn.freebsd.org/base/head@221432 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f